refactor(dashboard): address PR review — scope scroll fix, dedup, accurate counts
- scroll gutter: move the widget scroll-padding rule out of the shared UI
  globals into the dashboard-only stylesheet (scoped to .sero-dashboard) so it
  can't leak into full plugin views; drop scrollbar-gutter and keep a right
  padding, which is correct on every OS with no double spacing.
- orchestrator: extract useStateDir/dirOf and consume the shared
  useOrchestratorIndex hook in OrchestratorApp, removing the duplicated index
  path + watch logic.
- attention widget: count real pending items (questions + suggestions) in the
  header instead of collapsed display rows, matching LoopsWidget's "Needs you".
- time formatting: add a single relativeTime to @sero-ai/common (accepts
  number | string | Date) and remove the per-plugin copies; git commit feed
  now reads "5m ago" to match the web plugin's established wording.
