--- plugins/sero-git-plugin/ui/lib/format.ts
@@ -1,16 +1,0 @@
-/**
- * Value formatting shared by the git dashboard widgets.
- */
-
-/** Compact relative time from an ISO date: "now", "5m", "2h", "3d". */
-export function relativeTime(iso: string): string {
-  const then = new Date(iso).getTime();
-  if (Number.isNaN(then)) return '';
-  const sec = Math.round((Date.now() - then) / 1000);
-  if (sec < 60) return 'now';
-  const min = Math.round(sec / 60);
-  if (min < 60) return `${min}m`;
-  const hr = Math.round(min / 60);
-  if (hr < 24) return `${hr}h`;
-  return `${Math.round(hr / 24)}d`;
-}
