+++ plugins/sero-web-plugin/ui/lib/format.ts
@@ -1,16 +1,7 @@
 // lib/format.ts — Shared formatting helpers for the web UI.
 
-/** Format a timestamp as a relative time string (e.g. "2m ago"). */
-export function relativeTime(timestamp: number): string {
-  const seconds = Math.floor((Date.now() - timestamp) / 1000);
-  if (seconds < 60) return 'just now';
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes}m ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  return `${days}d ago`;
-}
+// Relative time ("2m ago") is shared across Sero UIs — see @sero-ai/common.
+export { relativeTime } from '@sero-ai/common';
 
 /** Extract the domain from a URL string. */
 export function extractDomain(url: string): string {
